pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labc"
description = 'This is the fives iteration of the lab-commander middleware.'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Meier, Moritz", email = "mome@uni-bremen.de" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "python-daemon",
  "toml",
  "appdirs",
  "psutil",
  "setproctitle",
  "paramiko",
]
dynamic = ["version"]

[project.optional-dependencies]
color = [
  "pygments",
  "coloredlogs",
  "click-help-colors",
]

[project.urls]
Documentation = "https://gitlab.csl.uni-bremen.de/labor/labc#readme"
Issues = "https://gitlab.csl.uni-bremen.de/labor/labc/issues"
Source = "https://gitlab.csl.uni-bremen.de/labor/labc"

[project.scripts]
labc = "labc.cli:labc"

[tool.hatch.version]
path = "labc/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=labc --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "labc/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
